# gaggiuino-local-profiler/openapi.yaml
openapi: 3.0.3
info:
  title: GLP REST API
  version: 2.0.0
  description: >
    REST API for the **Gaggiuino Local Profiler** Home Assistant add-on.
    Provides shot storage, coffee library, maintenance tracking, a barista
    orders system, and live machine data.
  contact:
    url: https://github.com/mxkissnr/gaggiuino-local-profiler
servers:
  - url: http://localhost:8099
    description: Direct port (integration, cards)
  - url: /api/hassio_ingress/gaggiuino_local_profiler
    description: HA Ingress (browser — auth via HA session, no token needed)
security:
  - tokenAuth: []
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: X-GLP-Token
      description: >
        64-char hex token, auto-generated on first start, stored in
        `/data/api_token.txt`. Retrieve from `GET /api/status`.
        Requests via HA Ingress bypass this check (HA session is sufficient).
        `GET /api/status` is always public.
  schemas:
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, example: true }
...
    MenuItem:
      type: object
      required: [id, name, emoji]
      properties:
        id:    { type: string, example: "m_1716000000000" }
        name:  { type: string, maxLength: 200 }
        emoji: { type: string, maxLength: 10 }
    Order:
      type: object
      required: [id, createdAt, customer, item, status]
      properties:
        id:            { type: string, example: "ord_1716000000000_ab12" }
        createdAt:     { type: integer, description: Unix ms }
        customer:      { type: string, maxLength: 50 }
        haUserId:      { type: string, maxLength: 100 }
        item:          { type: string, description: Menu item name }
        note:          { type: string, maxLength: 200 }
        status:
          type: string
          enum: [pending, accepted, done, declined]
        eta:           { type: integer, description: Minutes, nullable: true }
        acceptedAt:    { type: integer, description: Unix ms, nullable: true }
        completedAt:   { type: integer, description: Unix ms, nullable: true }
        declineReason: { type: string, nullable: true }
        shotId:        { type: integer, description: "Last shot ID when order completed — resolved from the order's own machineId (#326), not the global latest shot", nullable: true }
        machine:       { type: string, maxLength: 100, nullable: true, description: "Optional machine name/slug target as supplied by the client (#317/glp-order-card #29)" }
        machineId:     { type: integer, nullable: true, description: "machine's registry id, resolved server-side from `machine` (#326) — always set on orders placed since #326, falls back to the default machine when `machine` is unset/unmatched" }
        beanId:        { type: integer, nullable: true, description: "Library bean id for stable order-to-bean attribution (#563, glp-order-card #35). Resolved server-side against the library's actual beans — a stale/unknown id becomes null." }
    OrdersSettings:
      type: object
      properties:
        enabled: { type: boolean }
...
paths:
...
  /api/orders:
    get:
      tags: [Orders]
      summary: List orders (most recent first, max 100)
      description: >
        Requires `enable_orders: true` in add-on config, otherwise returns 404.
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, accepted, done, declined] }
        - name: machine
          in: query
          schema: { type: integer }
          description: "Scope to one machine's orders by its registry id (#326). Omitted = all machines."
      responses:
        "200":
          description: Array of orders
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Order' } }
        "404": { description: Orders feature not enabled }
    post:
      tags: [Orders]
      summary: Place a new order
      description: >
        `haUserId` should be supplied by the glp-integration proxy via the
        `X-GLP-HA-User-ID` header (injected from the authenticated HA session).
        The header value takes precedence over the body field.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [item, customer]
              properties:
                item:     { type: string, description: Must match a menu item name }
                customer: { type: string, maxLength: 50 }
                note:     { type: string, maxLength: 200 }
                haUserId: { type: string, maxLength: 100, description: Overridden by X-GLP-HA-User-ID header if present }
                machine:  { type: string, maxLength: 100, description: "Optional machine name/slug target (#317/glp-order-card #29). Resolved server-side (#326) into machineId against the machine registry — falls back to the default machine when unset or unmatched." }
                beanId:   { type: integer, description: "Optional library bean id for stable order-to-bean attribution (#563, glp-order-card #35). Resolved server-side against the library's actual beans — a stale/unknown id becomes null rather than failing order placement." }
      responses:
        "200": { description: Created order, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        "400": { description: Validation error }
        "429": { description: Rate limit exceeded }
        "503": { description: Orders disabled }
  /api/orders/mine:
    get:
      tags: [Orders]
      summary: Orders placed by a specific HA user (max 10, most recent first)
      parameters:
        - name: haUserId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Array of orders
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Order' } }
        "400": { description: haUserId required }
  /api/orders/{id}/accept:
    post:
      tags: [Orders]
      summary: Accept a pending order
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                eta: { type: integer, minimum: 1, maximum: 60, description: Estimated minutes, default 5 }
      responses:
        "200": { description: Updated order, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        "400": { description: Order not in pending state }
        "404": { description: Not found }
  /api/orders/{id}/complete:
    post:
      tags: [Orders]
      summary: Mark an order as done (links the most recent shot)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Updated order, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        "404": { description: Not found }
  /api/orders/{id}/decline:
    post:
      tags: [Orders]
      summary: Decline a pending or accepted order
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, maxLength: 200 }
      responses:
        "200": { description: Updated order, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        "400": { description: Cannot decline in current state }
        "404": { description: Not found }
  /api/orders/{id}:
    delete:
      tags: [Orders]
      summary: Delete a completed (done/declined) order from history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
        "400": { description: Can only delete completed orders }
        "404": { description: Not found }
...
  /api/orders/menu:
    get:
      tags: [Orders]
      summary: Get the current menu
      responses:
        "200":
          description: Array of menu items
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/MenuItem' } }
    post:
      tags: [Orders]
      summary: Add a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:  { type: string, maxLength: 200 }
                emoji: { type: string, maxLength: 10, default: "☕" }
      responses:
        "200": { description: Created item, content: { application/json: { schema: { $ref: '#/components/schemas/MenuItem' } } } }
        "400": { description: name required }
  /api/orders/menu/{id}:
    put:
      tags: [Orders]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:  { type: string, maxLength: 200 }
                emoji: { type: string, maxLength: 10 }
      responses:
        "200": { description: Updated item, content: { application/json: { schema: { $ref: '#/components/schemas/MenuItem' } } } }
        "404": { description: Not found }
    delete:
      tags: [Orders]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
        "404": { description: Not found }
  /api/orders/settings:
    get:
      tags: [Orders]
      summary: Get orders feature settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrdersSettings' }
    post:
      tags: [Orders]
      summary: Enable or disable the orders feature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        "200": { description: Updated settings, content: { application/json: { schema: { $ref: '#/components/schemas/OrdersSettings' } } } }
        "400": { description: enabled (boolean) required }
...